task-app/requirements.txt
Flask
Flask-SQLAlchemy
pymysql
gunicorn
python-dotenv
jsonify
Flask-JSON
flask-cors